update develop branch

+++ GramAddict/__main__.py
@@ -1,9 +1,9 @@
 import argparse
 from os import getcwd, path
 
+from GramAddict import __version__
 from GramAddict.core.bot_flow import start_bot
 from GramAddict.core.download_from_github import download_from_github
-from GramAddict.version import __version__
 
 
 def cmd_init(args):
@@ -26,6 +26,12 @@ def cmd_init(args):
                 )
             else:
                 print(f"'accounts/{username}' folder already exists, skip.")
+                continue
+            with open(f"./accounts/{username}/config.yml", "r+", encoding="utf-8") as f:
+                config = f.read()
+                f.seek(0)
+                config_fixed = config.replace("myusername", username)
+                f.write(config_fixed)
     else:
         print("You have to provide at last one account name..")
 
@@ -42,11 +48,15 @@ def cmd_dump(args):
     import uiautomator2 as u2
     from colorama import Fore, Style
 
-    os.popen("adb shell pkill atx-agent").close()
-    d = u2.connect()
+    if not args.no_kill:
+        os.popen("adb shell pkill atx-agent").close()
+    try:
+        d = u2.connect(args.device)
+    except RuntimeError as err:
+        raise SystemExit(err)
 
-    def dump_hierarchy(d, path):
-        xml_dump = d.dump_hierarchy()
+    def dump_hierarchy(device, path):
+        xml_dump = device.dump_hierarchy()
         with open(path, "w", encoding="utf-8") as outfile:
             outfile.write(xml_dump)
 
@@ -93,6 +103,19 @@ _commands = [
         action=cmd_dump,
         command="dump",
         help="dump current screen",
+        flags=[
+            dict(
+                args=["--device"],
+                nargs=None,
+                default=None,
+                help="provide the device name if more then one connected",
+            ),
+            dict(
+                args=["--no-kill"],
+                action="store_true",
+                help="don't kill the uia2 demon",
+            ),
+        ],
     ),
 ]
